🧹 [code health] Refactor PdfViewer to use custom hooks
