+++ apps/web/src/components/pdf-viewer-hooks/index.ts
@@ -1,0 +1,6 @@
+export * from "./types";
+export * from "./use-pdf-dimensions";
+export * from "./use-pdf-navigation";
+export * from "./use-pdf-search";
+export * from "./use-pdf-viewer";
+export * from "./use-pdf-zoom";
